% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergence_list.R
\name{divergence}
\alias{divergence}
\title{divergence}
\usage{
divergence(dvg = c("Chi2", "KL", "revKL", "Hellinger", "Alpha"), alpha = NULL)
}
\arguments{
\item{dvg}{Character. One of "\code{Chi2}", "\code{KL}", "\code{revKL}", "\code{Hellinger}", "\code{Alpha}" or "\code{user}". See 'Details'}

\item{alpha}{the value of the alpha parameter for the "Alpha" divergence}
}
\description{
divergence
}
\details{
The following divergence functions are considered:

Chi2 (\code{dvg=Chi2}):

\deqn{h(x)=x^2-1}

Kullback-Leibler (\code{dvg=KL}):

\deqn{h(x)=x \dot \log(x)}

reverse Kullback-Leibler (\code{dvg=revKL}):

\deqn{h(x)=-\log(x)}

Hellinger (\code{dvg=Hellinger}):

\deqn{h(x)=(1 - x) ^ {-2}}.

Alpha (\code{dvg=Alpha}):

\deqn{h(x)=\frac{x ^ \alpha - \alpha * (x - 1) - 1}{\alpha * (\alpha - 1)}}
}
\examples{
# example code




}
\author{
Pietro Millossovich
}
