% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_f_divergence.R
\name{stress_mean_div}
\alias{stress_mean_div}
\title{stress_mean_div}
\usage{
stress_mean_div(
  x,
  f = function(x) x,
  k = 1,
  m,
  div = c("Chi2", "KL", "Hellinger", "Alpha", "Triangular", "Jeffrey", "user"),
  inv.div = NULL,
  d.div = NULL,
  d.inv = NULL,
  p = rep(1/length(x), length(x)),
  alpha = NULL,
  normalise = TRUE,
  show = FALSE,
  names = NULL,
  start = NULL,
  sumRN = FALSE,
  ...
)
}
\arguments{
\item{x}{A data frame, matrix or vector +++++++++++++++++}

\item{f}{A function that, applied to \code{x}, constitute the moment constraints.}

\item{k}{A vector indicating which columns of \code{x} the function \code{f} operates on. By default, the first columnn is selected and the identity function is applied.}

\item{m}{Numeric, the stressed moments of \code{f(x)}. Must be in the
range of \code{f(x)}.}

\item{div}{Character. One of "Chi2", "KL", "Hellinger", "Alpha", "Triangular", "Jeffrey" or "user". When a user specified divergence is chosen, the additional parameters "inv.div" (inverse of the divergence function) and "d.div" (derivative of the divergence function) must be passed. For the "Alpha" divergence, the numeric parameter "alpha" must be provided (when alpha is 1 or 2 the KL, respectively the Chi2, divergence is used). +++++++++++++++++++ADD EQUATIONS FOR THE DIVERGENCES+++++++++++++++++}

\item{inv.div}{A function specifying the inverse of the divergence function, when a user divergence function is used.}

\item{d.div}{A function specifying the derivative of the divergence function, when a user divergence function is used.}

\item{d.inv}{An optional function specifying the derivative of the inverse of the divergence function, when a user divergence function is chosen.}

\item{p}{Numeric. A set of optional nonnegative scenario weights specifying the baseline model.}

\item{alpha}{Numeric. The 'alpha' parameter when the Alpha divergence is used.}

\item{normalise}{Logical. If true, values of \code{f(x)} are linearly scaled to the unit interval.}

\item{show}{Logical. If true, print the result of the call to \code{\link[nleqslv]{nleqslv}}.}

\item{names}{++++++++++++}

\item{start}{A numeric vector with two elements, the starting values for the coefficients lambda1 and lambda2. Defaults to d.div(1) and 0 respectively, guaranteeing that the initial set of scenario weights}

\item{sumRN}{Logical. If true, the scenario weights are normalized so as to average to 1.}

\item{...}{Additional arguments to be passed to \code{\link[nleqslv]{nleqslv}}.}
}
\value{
A named list
}
\description{
Provides weights on simulated scenarios from a baseline stochastic model, such that a stressed model component (random variable) fulfill a moment constraint. Scenario weights are selected by constrained minimisation of a selected divergence to baseline model.
}
\seealso{
See \code{\link{stress_moment}} for a more flexible function allowing to perform multiple joint stresses under the KL divergnce.
}
